fix embedding function interface
