+++ chromadb/utils/embedding_functions/cohere_embedding_function.py
@@ -1,8 +1,4 @@
-from chromadb.utils.embedding_functions.embedding_function import (
-    EmbeddingFunction,
-    Space,
-)
-from chromadb.api.types import Embeddings, Documents
+from chromadb.api.types import Embeddings, Documents, EmbeddingFunction, Space
 from typing import List, Dict, Any, Optional
 import os
 import numpy as np
