+++ chromadb/utils/embedding_functions/open_clip_embedding_function.py
@@ -1,7 +1,4 @@
-from chromadb.utils.embedding_functions.embedding_function import (
-    EmbeddingFunction,
-    Space,
-)
+from chromadb.api.types import EmbeddingFunction, Space
 from chromadb.utils.embedding_functions.schemas import validate_config
 from chromadb.api.types import (
     Document,
